requirements.txt
Django==1.10
django-crispy-forms==1.6.1
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-storages==1.6.3
djangorestframework==3.6.3
Markdown==2.6.8
markdown2==2.3.4
pytz==2017.2
